biscuit-python: bump pyo3 to 0.23.1
This is mostly just following the upgrade guide here: https://pyo3.rs/v0.22.0/migration.html

I chose to not go directly to 0.23.1 because (selfishly) we're on 0.22.6 and
don't want to wait on our upgrade.  Normally these python bindings would be
independent of what other modules in our codebase are doing, but in this case
I'm trying to build a single rust package with biscuit-python + bindings to our
own layer on top of biscuits, so they need to use the same pyo3 version.

Test Plan: